requirements.txt
pip
torch
transformers
smolagents[litellm]
smolagents[transformers]
smolagents[openai]
smolagents[telemetry,toolkit]
ddgs
requests
duckduckgo-search
wikipedia
pandas
pillow
accelerate
num2words
quanto
bitsandbytes
optimum-quanto
pytest
pytest-mock
arize-phoenix

